fix(tangle-cloud): refactor earnings page into components and extract mappers
Extract earnings table into separate components, add token presentation utils,
and move GraphQL data mapping into dedicated mapper files for developer payments
and reward claims.
